Rental system with Database
This is a Java GUI system, in the add tab, there are two options, Add a customer and Add a rental. Add customer, takes the customer contact information and adds it to the database. I used Regex to control user's input.

When user clicks on "Click to Submit," they are added to the database.

On the Database tab, when user clicks on "Write Database Tables," the database is written in the system propmpt.

On the exit tab, "Calculate Charges," calculates and prints all custmer charges in a dialog box.